Fixed all Issues
More question added. All the questions are still being fetched from Open Trivia Database.

A trivia consist of 12 questions.

Implemented a method to check when there are no more questions left so the app doesn't crash. Upon completing the trivia an pop is initialised which congratulates the user and shows them the number of answers marked correct.

At pop up they can also choose to retake the trivia with a new set of questions.

While the question is being fetched a custom circular progress indicator is shown.

The debug banner was also removed.